Change docUnitId to String, add valid check
RISDEV-105

+++ src/main/java/de/bund/digitalservice/ris/controller/DocUnitController.java
@@ -24,6 +24,17 @@ public class DocUnitController {
     this.service = service;
   }
 
+  // only general check, the service will check if such a doc unit exists
+  private boolean isInvalidId(String id) {
+    // this validation has to change once we use non-integer Ids
+    try {
+      Integer.parseInt(id);
+      return false;
+    } catch (NumberFormatException e) {
+      return true;
+    }
+  }
+
   @PostMapping(value = "")
   public Mono<ResponseEntity<DocUnit>> generateNewDocUnit() {
 
@@ -32,16 +43,21 @@ public class DocUnitController {
 
   @PutMapping(value = "/{id}/file")
   public Mono<ResponseEntity<DocUnit>> attachFileToDocUnit(
-      @PathVariable int id,
+      @PathVariable String id,
       @RequestBody Flux<ByteBuffer> byteBufferFlux,
       @RequestHeader HttpHeaders httpHeaders) {
+    if (isInvalidId(id)) {
+      return Mono.just(ResponseEntity.unprocessableEntity().body(DocUnit.EMPTY));
+    }
 
     return service.attachFileToDocUnit(id, byteBufferFlux, httpHeaders);
   }
 
   @DeleteMapping(value = "/{id}/file")
-  public Mono<ResponseEntity<DocUnit>> removeFileFromDocUnit(@PathVariable int id) {
-
+  public Mono<ResponseEntity<DocUnit>> removeFileFromDocUnit(@PathVariable String id) {
+    if (isInvalidId(id)) {
+      return Mono.just(ResponseEntity.unprocessableEntity().body(DocUnit.EMPTY));
+    }
     return service.removeFileFromDocUnit(id);
   }
 
@@ -53,15 +69,18 @@ public class DocUnitController {
   }
 
   @GetMapping(value = "/{id}")
-  public Mono<ResponseEntity<DocUnit>> getById(@PathVariable int id) {
+  public Mono<ResponseEntity<DocUnit>> getById(@PathVariable String id) {
+    if (isInvalidId(id)) {
+      return Mono.just(ResponseEntity.unprocessableEntity().body(DocUnit.EMPTY));
+    }
     return service.getById(id);
   }
 
   @PutMapping(value = "/{id}/docx", consumes = MediaType.APPLICATION_JSON_VALUE)
   public Mono<ResponseEntity<DocUnit>> updateById(
-      @PathVariable int id, @RequestBody DocUnit docUnit) {
-    if (id != docUnit.getId()) {
-      return Mono.just(ResponseEntity.internalServerError().body(DocUnit.EMPTY));
+      @PathVariable String id, @RequestBody DocUnit docUnit) {
+    if (isInvalidId(id) || !Integer.valueOf(id).equals(docUnit.getId())) {
+      return Mono.just(ResponseEntity.unprocessableEntity().body(DocUnit.EMPTY));
     }
     return service.updateDocUnit(docUnit);
   }
